Add minimal CLI program.
Added a simple binary that uses the interpreter lib. It takes a path as
a single positional argument, and loads any file that exists at that
path into memory. This data is then passed to the interpreter, which
attempts to run it as a chip-8 program. No display or sound or telemetry
of any kind are currently included.

Tested on a simple test chip-8 program and observed the interpreter run
it without encountering errors.